requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
python-keycloak==5.8.1
email-validator==2.2.0
prometheus-fastapi-instrumentator==6.1.0

# File storage dependencies
minio==7.2.0
Pillow==10.0.0
python-magic==0.4.27
python-multipart==0.0.6

aiohttp==3.13.2